# docs/services/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CORTX Gateway
  version: 0.1.0
paths:
...
  /fedsuite/{path}:
    post:
      tags:
      - FedSuite
      - FedSuite Proxy
      summary: Proxy All
      description: Proxy all other requests with JWT validation
      operationId: proxy_all_fedsuite__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - FedSuite
      - FedSuite Proxy
      summary: Proxy All
      description: Proxy all other requests with JWT validation
      operationId: proxy_all_fedsuite__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - FedSuite
      - FedSuite Proxy
      summary: Proxy All
      description: Proxy all other requests with JWT validation
      operationId: proxy_all_fedsuite__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - FedSuite
      - FedSuite Proxy
      summary: Proxy All
      description: Proxy all other requests with JWT validation
      operationId: proxy_all_fedsuite__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError